Make dyntopo detail size maximum smaller (more than 40 pixels should not be needed according to mapping in code) and change the property to float.

+++ source/blender/makesdna/DNA_scene_types.h
@@ -848,7 +848,7 @@ typedef struct Sculpt {
 	int radial_symm[3];
 
 	/* Maximum edge length for dynamic topology sculpting (in pixels) */
-	int detail_size;
+	float detail_size;
 
 	/* Direction used for SCULPT_OT_symmetrize operator */
 	int symmetrize_direction;
